requirements_ui.txt
pandas==1.3.4
pyArango==1.3.5
nltk==3.6.2
scikit-learn==0.24.2
python-dotenv==0.18.0
regex==2021.7.6
Flask==2.0.2
Flask-Cors==3.0.10
py-espeak-ng==0.1.8
Flask-RESTful==0.3.9
boto3==1.18.36
pytorch-transformers==1.2.0
torch==1.9.0